Make the AST explicit as an AST
