+++ lib/src/main/java/com/lucaskjaerozhang/wikitext_parser/Parser.java
@@ -1,6 +1,6 @@
 package com.lucaskjaerozhang.wikitext_parser;
 
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextElement;
+import com.lucaskjaerozhang.wikitext_parser.ast.base.WikiTextElement;
 import com.lucaskjaerozhang.wikitext_parser.parse.SetupParse;
 
 public class Parser {
