+++ lib/src/main/java/com/lucaskjaerozhang/wikitext_parser/ast/format/Italic.java
@@ -1,0 +1,19 @@
+package com.lucaskjaerozhang.wikitext_parser.ast.format;
+
+import com.lucaskjaerozhang.wikitext_parser.ast.base.WikiTextElement;
+import com.lucaskjaerozhang.wikitext_parser.ast.base.WikiTextNode;
+import com.lucaskjaerozhang.wikitext_parser.ast.base.WikiTextParentNode;
+import java.util.List;
+
+public class Italic extends WikiTextParentNode implements WikiTextElement {
+  public static final String XML_TAG = "italic";
+
+  public Italic(List<WikiTextNode> content) {
+    super(content);
+  }
+
+  @Override
+  public String getXMLTag() {
+    return XML_TAG;
+  }
+}
