+++ lib/src/main/java/com/lucaskjaerozhang/wikitext_parser/ast/link/WikiLinkNamespaceComponentType.java
@@ -1,0 +1,8 @@
+package com.lucaskjaerozhang.wikitext_parser.ast.link;
+
+public enum WikiLinkNamespaceComponentType {
+  CATEGORY,
+  LANGUAGE,
+  NAMESPACE,
+  WIKI,
+}
