+++ lib/src/main/java/com/lucaskjaerozhang/wikitext_parser/ast/list/ListType.java
@@ -1,0 +1,18 @@
+package com.lucaskjaerozhang.wikitext_parser.ast.list;
+
+public enum ListType {
+  UNORDERED("unordered"),
+  ORDERED("ordered"),
+  DESCRIPTION("description");
+
+  private final String type;
+
+  ListType(String type) {
+    this.type = type;
+  }
+
+  @Override
+  public String toString() {
+    return type;
+  }
+}
