--- lib/src/main/java/com/lucaskjaerozhang/wikitext_parser/objects/format/Italic.java
@@ -1,19 +1,0 @@
-package com.lucaskjaerozhang.wikitext_parser.objects.format;
-
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextElement;
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextNode;
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextParentNode;
-import java.util.List;
-
-public class Italic extends WikiTextParentNode implements WikiTextElement {
-  public static final String XML_TAG = "italic";
-
-  public Italic(List<WikiTextNode> content) {
-    super(content);
-  }
-
-  @Override
-  public String getXMLTag() {
-    return XML_TAG;
-  }
-}
