--- lib/src/main/java/com/lucaskjaerozhang/wikitext_parser/objects/link/CategoryLink.java
@@ -1,33 +1,0 @@
-package com.lucaskjaerozhang.wikitext_parser.objects.link;
-
-import java.util.Set;
-
-/** Categories are just a special case of a link. */
-public class CategoryLink extends WikiLink {
-  public static final String XML_TAG = "category";
-  private final String category;
-  private final boolean visible;
-
-  public CategoryLink(WikiLinkTarget linkTarget, String linkText, boolean visible) {
-    super(linkTarget, linkText);
-    this.category = linkTarget.wholeLink();
-    this.visible = visible;
-  }
-
-  @Override
-  public String getXMLTag() {
-    return XML_TAG;
-  }
-
-  @Override
-  public Set<String> getCategories() {
-    return Set.of(category);
-  }
-
-  @Override
-  public String toXML() {
-    // Category links can be invisible.
-    // This is done to place an article in a category.
-    return visible ? super.toXML() : "";
-  }
-}
