--- lib/src/main/java/com/lucaskjaerozhang/wikitext_parser/objects/link/WikiLinkNamespaceComponentType.java
@@ -1,8 +1,0 @@
-package com.lucaskjaerozhang.wikitext_parser.objects.link;
-
-public enum WikiLinkNamespaceComponentType {
-  CATEGORY,
-  LANGUAGE,
-  NAMESPACE,
-  WIKI,
-}
