--- lib/src/main/java/com/lucaskjaerozhang/wikitext_parser/objects/list/ListItem.java
@@ -1,40 +1,0 @@
-package com.lucaskjaerozhang.wikitext_parser.objects.list;
-
-import com.lucaskjaerozhang.wikitext_parser.objects.base.NodeAttribute;
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextElement;
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextNode;
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextParentNode;
-import java.util.List;
-import java.util.Optional;
-
-public class ListItem extends WikiTextParentNode implements WikiTextElement {
-  public static final String XML_TAG = "listItem";
-  public static final String LEVEL_ATTRIBUTE = "level";
-
-  private final Optional<Integer> level;
-
-  public ListItem(Optional<Integer> level, WikiTextNode content) {
-    super(List.of(content));
-    this.level = level;
-  }
-
-  public ListItem(Optional<Integer> level, List<WikiTextNode> content) {
-    super(content);
-    this.level = level;
-  }
-
-  @Override
-  public String getXMLTag() {
-    return XML_TAG;
-  }
-
-  @Override
-  public List<NodeAttribute> getAttributes() {
-    if (level.isEmpty()) return List.of();
-    return List.of(new NodeAttribute(LEVEL_ATTRIBUTE, level.get().toString(), false));
-  }
-
-  public Integer getLevel() {
-    return level.orElse(1);
-  }
-}
