--- lib/src/main/java/com/lucaskjaerozhang/wikitext_parser/objects/sections/HorizontalRule.java
@@ -1,12 +1,0 @@
-package com.lucaskjaerozhang.wikitext_parser.objects.sections;
-
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextLeafNode;
-
-public class HorizontalRule extends WikiTextLeafNode {
-  public static final String XML_TAG = "horizontalRule";
-
-  @Override
-  public String getXMLTag() {
-    return XML_TAG;
-  }
-}
