+++ lib/src/main/java/com/lucaskjaerozhang/wikitext_parser/parse/WikitextVisitor.java
@@ -2,28 +2,28 @@ package com.lucaskjaerozhang.wikitext_parser.parse;
 
 import com.lucaskjaerozhang.wikitext_parser.grammar.WikiTextBaseVisitor;
 import com.lucaskjaerozhang.wikitext_parser.grammar.WikiTextParser;
-import com.lucaskjaerozhang.wikitext_parser.objects.base.NodeAttribute;
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextElement;
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextNode;
-import com.lucaskjaerozhang.wikitext_parser.objects.base.WikiTextParentNode;
-import com.lucaskjaerozhang.wikitext_parser.objects.format.Bold;
-import com.lucaskjaerozhang.wikitext_parser.objects.format.Italic;
-import com.lucaskjaerozhang.wikitext_parser.objects.layout.IndentedBlock;
-import com.lucaskjaerozhang.wikitext_parser.objects.layout.XMLContainerElement;
-import com.lucaskjaerozhang.wikitext_parser.objects.layout.XMLStandaloneElement;
-import com.lucaskjaerozhang.wikitext_parser.objects.link.CategoryLink;
-import com.lucaskjaerozhang.wikitext_parser.objects.link.WikiLink;
-import com.lucaskjaerozhang.wikitext_parser.objects.link.WikiLinkNamespaceComponent;
-import com.lucaskjaerozhang.wikitext_parser.objects.link.WikiLinkTarget;
-import com.lucaskjaerozhang.wikitext_parser.objects.list.ListItem;
-import com.lucaskjaerozhang.wikitext_parser.objects.list.ListType;
-import com.lucaskjaerozhang.wikitext_parser.objects.list.WikiTextList;
-import com.lucaskjaerozhang.wikitext_parser.objects.root.Article;
-import com.lucaskjaerozhang.wikitext_parser.objects.root.CategoryList;
-import com.lucaskjaerozhang.wikitext_parser.objects.root.Redirect;
-import com.lucaskjaerozhang.wikitext_parser.objects.sections.HorizontalRule;
-import com.lucaskjaerozhang.wikitext_parser.objects.sections.Section;
-import com.lucaskjaerozhang.wikitext_parser.objects.sections.Text;
+import com.lucaskjaerozhang.wikitext_parser.ast.base.NodeAttribute;
+import com.lucaskjaerozhang.wikitext_parser.ast.base.WikiTextElement;
+import com.lucaskjaerozhang.wikitext_parser.ast.base.WikiTextNode;
+import com.lucaskjaerozhang.wikitext_parser.ast.base.WikiTextParentNode;
+import com.lucaskjaerozhang.wikitext_parser.ast.format.Bold;
+import com.lucaskjaerozhang.wikitext_parser.ast.format.Italic;
+import com.lucaskjaerozhang.wikitext_parser.ast.layout.IndentedBlock;
+import com.lucaskjaerozhang.wikitext_parser.ast.layout.XMLContainerElement;
+import com.lucaskjaerozhang.wikitext_parser.ast.layout.XMLStandaloneElement;
+import com.lucaskjaerozhang.wikitext_parser.ast.link.CategoryLink;
+import com.lucaskjaerozhang.wikitext_parser.ast.link.WikiLink;
+import com.lucaskjaerozhang.wikitext_parser.ast.link.WikiLinkNamespaceComponent;
+import com.lucaskjaerozhang.wikitext_parser.ast.link.WikiLinkTarget;
+import com.lucaskjaerozhang.wikitext_parser.ast.list.ListItem;
+import com.lucaskjaerozhang.wikitext_parser.ast.list.ListType;
+import com.lucaskjaerozhang.wikitext_parser.ast.list.WikiTextList;
+import com.lucaskjaerozhang.wikitext_parser.ast.root.Article;
+import com.lucaskjaerozhang.wikitext_parser.ast.root.CategoryList;
+import com.lucaskjaerozhang.wikitext_parser.ast.root.Redirect;
+import com.lucaskjaerozhang.wikitext_parser.ast.sections.HorizontalRule;
+import com.lucaskjaerozhang.wikitext_parser.ast.sections.Section;
+import com.lucaskjaerozhang.wikitext_parser.ast.sections.Text;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
